requirements.txt
numpy
matplotlib
seaborn
opencv-python
tensorflow
scikit-learn
Pillow
jupyter